requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.30
python-docx==1.1.2
openpyxl==3.1.2
